Added insertion sort algo to sort arrays of integers.
This C++ program illustrates our conventions for implementing basic array sorts
The main program function is a driver that initializes an array of integer
(either with random values or from standard input), calls a sort function to
sort that array, then prints out the ordered result.

Templates allow the implementation to be used to sort items of data type for
which comparison and assignment are defined.

The sort function here is a version of insertion sort. It uses a template
function that compares 2 items and exchanges them if necessary to make the
second not less ( < ) than the first.

we can change the driver to sort any type of data for which operator "<" is
defined without modifying sort at all.

Similar inplemention is C is implemented as well.

Signed-off-by: Raja Mallik <raja.mallik@alumni.iitm.ac.in>